Fix handling of Jupyterlab version
