+++ jupyterlab/package.json
@@ -2,7 +2,7 @@
   "private": true,
   "name": "@jupyterlab/application-top",
   "scripts": {
-    "build": "webpack"
+    "build": "node build.js && webpack"
   },
   "dependencies": {
     "@jupyterlab/about-extension": "^0.3.1",
@@ -76,7 +76,9 @@
       "@jupyterlab/tabmanager-extension",
       "@jupyterlab/terminal-extension",
       "@jupyterlab/tooltip-extension"
-    ]
+    ],
+    "version": "0.20.1",
+    "gitDescription": "v0.18.0-0-937-g35b86e85"
   },
   "version": "0.1.2"
 }
